Add some more clothing and weapons
Surcoats, half-tabards (pants), scapulars: Intended to be fodder for decorating.  Half-tabards are meant to be the item of decorative lower-body-wear that is ubiquitous in fantasy and apparently nonexistant in real life.  Zelda, in particular, tends to wear one decorated with Hyrule and sometimes Sheikah symbols on it.

Underwear: I didn't like that *every* civ's members were described as wearing loincloths.  "Briefs" may not be a good substitution, though :/

Katana: Good sword for Sheikah Heritics

Bandana: Pirate wear

Shalwar: Baggy pants